+++ lib/src/core/vec.dart
@@ -8,6 +8,7 @@ import 'package:ffi/ffi.dart';
 
 import '../g/types.g.dart' as cvg;
 import 'base.dart';
+import 'float16.dart';
 
 abstract class Vec<N extends ffi.Struct, T> with IterableMixin<T> implements ffi.Finalizable {
   Vec.fromPointer(this.ptr);
@@ -664,54 +665,6 @@ class VecF16Iterator extends VecIterator<double> {
 
   @override
   double operator [](int idx) => float16(ref.ptr[idx]);
-}
-
-double float16(int w) {
-  final t = calloc<ffi.UnsignedInt>()..value = ((w & 0x7fff) << 13) + 0x38000000;
-  final sign = calloc<ffi.UnsignedInt>()..value = (w & 0x8000) << 16;
-  final e = calloc<ffi.UnsignedInt>()..value = w & 0x7c00;
-
-  final out = calloc<cvg.Cv32suf_C>()..ref.u = t.value + (1 << 23);
-  if (e.value >= 0x7c00) {
-    out.ref.u = t.value + 0x38000000;
-  } else {
-    if (e.value == 0) {
-      out.ref.f -= 6.103515625e-05;
-    } else {
-      out.ref.u = t.value;
-    }
-  }
-  out.ref.u |= sign.value;
-  final rval = out.ref.f;
-  calloc.free(out);
-  calloc.free(t);
-  calloc.free(sign);
-  calloc.free(e);
-  return rval;
-}
-
-int float16Inv(double x) {
-  final in_ = calloc<cvg.Cv32suf_C>()..ref.f = x;
-  final sign = calloc<ffi.UnsignedInt>()..value = in_.ref.u & 0x80000000;
-  final w = calloc<ffi.UnsignedInt>();
-  in_.ref.u ^= sign.value;
-  if (in_.ref.u > 0x47800000) {
-    w.value = in_.ref.u > 0x7f800000 ? 0x7e00 : 0x7c00;
-  } else {
-    if (in_.ref.u < 0x38800000) {
-      in_.ref.f += 0.5;
-      w.value = in_.ref.u - 0x3f000000;
-    } else {
-      final t = calloc<ffi.UnsignedInt>()..value = in_.ref.u + 0xc8000fff;
-      w.value = (t.value + ((in_.ref.u >> 13) & 1)) >> 13;
-      calloc.free(t);
-    }
-  }
-  final rval = w.value | (sign.value >> 16);
-  calloc.free(in_);
-  calloc.free(sign);
-  calloc.free(w);
-  return rval;
 }
 
 extension DoubleFp16Extension on double {
